Refactor repeat interval amount from Long to Int

+++ RemindMe/app/src/main/java/dev/shorthouse/remindme/domain/reminder/CompleteReminderUseCase.kt
@@ -65,7 +65,7 @@ class CompleteReminderUseCase @Inject constructor(
         val repeatInterval = reminder.repeatInterval ?: return reminder.startDateTime
 
         val repeatDuration = repeatInterval.unit.duration
-            .multipliedBy(repeatInterval.amount)
+            .multipliedBy(repeatInterval.amount.toLong())
             .toKotlinDuration()
 
         if (reminder.startDateTime.isAfter(ZonedDateTime.now())) {
